Cleanup codebase of dead code and unused imports

+++ backend/src/main/java/com/beaverbyte/financial_tracker_application/security/jwt/JwtUtils.java
@@ -104,7 +104,7 @@ public class JwtUtils {
 		return ResponseCookie
 				.from(name, value)
 				.path(path)
-				.maxAge(24 * 60 * 60)
+				.maxAge((long) 24 * 60 * 60)
 				.httpOnly(true).build();
 	}
